Refactoring following review - part 1.
